Fix issue with underscore `pick` of file properties
For some strange reason, underscore.js (currently on meteor v1.0.4) does not project the listed properties when using its `pick` function with a variable list of arguments for keys. Replacing the list of args with a single array (both listed as viable options in the underscore docs) fixes the issue in our current environment, and shouldn't affect other setups.